+++ AlgorithmsRanking/Controllers/ResearchesController.cs
@@ -24,13 +24,22 @@ namespace AlgorithmsRanking.Controllers
         [HttpGet]
         public async Task<IActionResult> Get()
         {
-            return Ok(await _db.GetResearchesAsync());
+            var items = await _db.GetResearchesAsync();
+
+            return Ok(items);
         }
 
         [HttpGet("{id:int}")]
         public async Task<IActionResult> Get(int id)
         {
-            return Ok(await _db.GetResearchAsync(id));
+            var model = await _db.GetResearchAsync(id);
+
+            if (model == null)
+            {
+                return NotFound(new ApiError("404", "Not Found", $"Исследование #{id} не найдено"));
+            }
+
+            return Ok(model);
         }
 
         [HttpGet("create")]
@@ -39,6 +48,7 @@ namespace AlgorithmsRanking.Controllers
             var model = new ResearchForm
             {
                 Id = 0,
+                Research = new Research(),
                 Init = new ResearchInitForm(),
                 Calculated = null,
                 Algorithms = await _db.GetAlgorithmsListItemsAsync(),
@@ -59,6 +69,11 @@ namespace AlgorithmsRanking.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody]ResearchInitForm model)
         {
+            if (model == null)
+            {
+                return BadRequest(new ApiError("400", "Null model", $"{typeof(ResearchInitForm)} cannot be null"));
+            }
+
             try
             {
                 model.CreatorId = 1; // TODO : заменить на реальный id!
@@ -74,7 +89,7 @@ namespace AlgorithmsRanking.Controllers
             }
             catch (Exception ex)
             {
-                return BadRequest(ex);
+                return BadRequest(new ApiError(ex));
             }
         }
 
@@ -91,21 +106,23 @@ namespace AlgorithmsRanking.Controllers
         {
             var item = await _db.GetResearchAsync(id);
 
+            if (item == null)
+            {
+                return NotFound(new ApiError("404", "Not Found", $"Исследование #{id} не найдено"));
+            }
+
             var model = new ResearchForm
             {
                 Id = item.Id,
+                Research = item,
                 Init = new ResearchInitForm
                 {
                     Name = item.Name,
                     Description = item.Description,
                     AlgorithmId = item.AlgorithmId,
-                    Algorithm = item.Algorithm,
                     DataSetId = item.DataSetId,
-                    DataSet = item.DataSet,
                     CreatorId = item.CreatorId,
-                    Creator = item.Creator,
                     ExecutorId = item.ExecutorId,
-                    Executor = item.Executor
                 },
                 Calculated = item.AccuracyRate.HasValue && item.EfficiencyRate.HasValue ?
                     new ResearchCalculatedForm
@@ -125,6 +142,16 @@ namespace AlgorithmsRanking.Controllers
         [HttpPut("{id:int}")]
         public async Task<IActionResult> Edit([FromRoute]int id, [FromBody]ResearchInitForm model)
         {
+            if (model == null)
+            {
+                return BadRequest(new ApiError("400", "Null model", $"{typeof(ResearchInitForm)} cannot be null"));
+            }
+
+            if ((await _db.GetResearchAsync(id)) == null)
+            {
+                return NotFound(new ApiError("404", "Not Found", $"Исследование #{id} не найдено"));
+            }
+
             try
             {
                 var result = await _db.UpdateResearchAsync(id, model);
@@ -138,13 +165,18 @@ namespace AlgorithmsRanking.Controllers
             }
             catch (Exception ex)
             {
-                return BadRequest(ex);
+                return BadRequest(new ApiError(ex));
             }
         }
 
         [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute]int id)
         {
+            if ((await _db.GetResearchAsync(id)) == null)
+            {
+                return NotFound(new ApiError("404", "Not Found", $"Исследование #{id} не найдено"));
+            }
+
             try
             {
                 await _db.RemoveResearchAsync(id);
@@ -153,7 +185,7 @@ namespace AlgorithmsRanking.Controllers
             }
             catch (Exception ex)
             {
-                return BadRequest(ex);
+                return BadRequest(new ApiError(ex));
             }
         }
 
@@ -166,20 +198,25 @@ namespace AlgorithmsRanking.Controllers
             }
             catch (Exception ex)
             {
-                return BadRequest(ex);
+                return BadRequest(new ApiError(ex));
             }
         }
 
         [HttpPost("{id:int}/execute")]
         public async Task<IActionResult> Execute([FromRoute]int id, [FromBody]ResearchCalculatedForm rates)
         {
+            if (rates == null)
+            {
+                return BadRequest(new ApiError("400", "Null model", $"Не заданы вычисленные параметры"));
+            }
+
             try
             {
                 return Ok(await _db.ExecuteResearchAsync(id, rates));
             }
             catch (Exception ex)
             {
-                return BadRequest(ex);
+                return BadRequest(new ApiError(ex));
             }
         }
 
@@ -192,7 +229,7 @@ namespace AlgorithmsRanking.Controllers
             }
             catch (Exception ex)
             {
-                return BadRequest(ex);
+                return BadRequest(new ApiError(ex));
             }
         }
 
@@ -205,7 +242,7 @@ namespace AlgorithmsRanking.Controllers
             }
             catch (Exception ex)
             {
-                return BadRequest(ex);
+                return BadRequest(new ApiError(ex));
             }
         }
     }
